+++ src/main/java/roomescape/auth/LoginInterceptor.java
@@ -6,7 +6,6 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.web.servlet.HandlerInterceptor;
 import roomescape.exception.UnauthorizedException;
-import roomescape.member.LoginMember;
 
 public class LoginInterceptor implements HandlerInterceptor {
