+++ src/main/java/roomescape/reservation/ReservationController.java
@@ -10,7 +10,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.AuthenticatedMember;
-import roomescape.member.LoginMember;
+import roomescape.auth.LoginMember;
 
 @RestController
 public class ReservationController {
